Create magicSquare
This was for our Java programming class. It is designed to read in a text file called input that contains some data about magic squares. Primarily the data was formatted so the first integer in the list defined the size of the array and then followed with integers to be put into the array. It was then required to sum the rows, columns, and diagonals to determine if it was a magic square or not.